Application is a pure abstract interface now
