+++ scorex-basics/src/main/scala/scorex/api/http/PeersApiRoute.scala
@@ -9,7 +9,7 @@ import akka.http.scaladsl.server.Route
 import akka.pattern.ask
 import io.swagger.annotations._
 import play.api.libs.json.{JsArray, JsString, Json}
-import scorex.app.Application
+import scorex.app.RunnableApplication
 import scorex.network.Handshake
 import scorex.network.NetworkController.ConnectTo
 import scorex.network.peer.{PeerInfo, PeerManager}
@@ -19,7 +19,7 @@ import scala.util.Try
 
 @Path("/peers")
 @Api(value = "/peers", description = "Get info about peers", position = 2)
-case class PeersApiRoute(override val application: Application)(implicit val context: ActorRefFactory)
+case class PeersApiRoute(override val application: RunnableApplication)(implicit val context: ActorRefFactory)
   extends ApiRoute {
 
   override lazy val route =
