+++ scorex-transaction/src/main/scala/scorex/api/http/AddressApiRoute.scala
@@ -9,7 +9,7 @@ import akka.http.scaladsl.server.Route
 import io.swagger.annotations._
 import play.api.libs.json._
 import scorex.account.{Account, PublicKeyAccount}
-import scorex.app.Application
+import scorex.app.RunnableApplication
 import scorex.crypto.EllipticCurveImpl
 import scorex.crypto.encode.Base58
 
@@ -17,7 +17,7 @@ import scala.util.{Failure, Success, Try}
 
 @Path("/addresses")
 @Api(value = "/addresses/", description = "Info about wallet's accounts and other calls about addresses")
-case class AddressApiRoute(override val application: Application)(implicit val context: ActorRefFactory)
+case class AddressApiRoute(override val application: RunnableApplication)(implicit val context: ActorRefFactory)
   extends ApiRoute with CommonTransactionApiFunctions {
 
   private val wallet = application.wallet
