+++ src/test/scala/scorex/lagonaki/server/LagonakiApplication.scala
@@ -3,7 +3,7 @@ package scorex.lagonaki.server
 import akka.actor.Props
 import com.typesafe.config.ConfigFactory
 import scorex.api.http._
-import scorex.app.{Application, ApplicationVersion}
+import scorex.app.{RunnableApplication, ApplicationVersion}
 import scorex.consensus.nxt.api.http.NxtConsensusApiRoute
 import scorex.lagonaki.mocks.ConsensusMock
 import scorex.network._
@@ -11,7 +11,7 @@ import scorex.transaction._
 
 import scala.reflect.runtime.universe._
 
-class LagonakiApplication(val settingsFilename: String) extends Application {
+class LagonakiApplication(val settingsFilename: String) extends RunnableApplication {
 
   override val applicationName = "lagonaki"
